requirements.txt
typer==0.4.0
requests==2.26.0
mysql-connector-python==8.0.27